E1: Added a bunch of comments, still stuck
Didn't try to change the "reset" for another implementation. Could've maybe tried to switch the animals (0-9) to memes, but I feel I wouldn't really think learn much aside knowing how to do something similar with different images (probably the effort is changing the meaning and context for the function and text). Learning to reset seems more important to me.